// src/models/systemparameters.ts
export interface SystemParametersModel extends BaseModel {
  [key: string]: any;
  recid?: number;
  service_charge?: number;
  background_color?: string;
  rddocnum?: string;
  rsdocnum?: string;
  tidocnum?: string;
  todocnum?: string;
  wsdocnum?: string;
  pcdocnum?: string;
  ordocnum?: string;
  posdocnum?: string;
  billnum?: string;
  seqnum?: string;
  voidnum?: string;
  ftphost?: string;
  ftpuser?: string;
  ftppword?: string;
  ftpport?: number;
  ftp?: number;
  robinson?: number;
  ayala?: number;
  megaworld?: number;
  sm?: number;
  naia?: number;
  ortigas?: number;
  vatrte?: number;
  localtax?: number;
  pathfile?: string;
  timestart?: string;
  timeend?: string;
  datestart?: Date;
  orderprintcount?: number;
  mw_fetchhrcde?: number;
  ticket_bysubclass?: boolean;
  enable_default_ticket_bysubclass?: boolean;
  takeout_scharge?: number;
  allow_carryover?: number;
  itemphoto_able?: number;
  billdocnum?: string;
  ordercde?: string;
  otc?: number;
  internal_param_inventory?: number;
  internal_param_customer?: number;
  posdelord?: number;
  cusdocnum?: string;
  orderdeldocnum?: string;
  itmclanum?: string;
  itmsubclanum?: string;
  memcnum?: string;
  locnum?: string;
  withtracc?: number;
  vdcdecemupd?: string;
  crdtypcemupd?: string;
  frrsncemupd?: string;
  pytpcemupd?: string;
  arcdecemupd?: string;
  itmclacemupd?: string;
  itmsbclacemupd?: string;
  mdcdecemupd?: string;
  itmcemupd?: string;
  activateautotransfer?: number;
  transferinterval?: number;
  serverprotocol?: string;
  serversite?: string;
  serveripaddress?: string;
  serverport?: string;
  serverhost?: string;
  serverusername?: string;
  serverpassword?: string;
  serverfileport?: string;
  trnsfrmod?: string;
  itmcdedocnum?: string;
  discdedocnum?: string;
  reasondocnum?: string;
  holidaydocnum?: string;
  combodocnum?: string;
  vreasondocnum?: string;
  tabledocnum?: string;
  internal_param?: number;
  no_dineout?: number;
  auth_itm_remove?: number;
  dinein_scharge?: number;
  pending_checklist?: number;
  accpac?: number;
  send_to_kitchen?: number;
  sticker_printer?: number;
  is_transferring?: number;
  bypass_trnlock?: boolean;
  internal_param_selfservice?: number;
  selfservice_zread_time?: number;
  selfservice_firstzread_trncount?: number;
  allow_otherlanguage?: number;
  swipe_auth?: number;
  auth_cancel_tran?: number | boolean;
  email_report?: number;
  auth_void_tran?: number;
  auth_reprintvoid_tran?: number;
  auth_reprint_tran?: number;
  auth_recall_tran?: number;
  auth_add_disc?: number;
  auth_free_itm?: number;
  auth_free_tran?: number;
  auth_report?: number;
  bypass_hookup_resend?: number;
  ej_pathfile?: string;
  dinetypenum?: string;
  warehousenum?: string;
  pricelistnum?: string;
  cusnum?: string;
  multibranch?: number;
  multiwar?: number;
  manual_dinetype?: number;
  auth_prc_change?: number;
  sync_last_update?: Date;
  sticker_path?: string;
  dbbackup_pathfile?: string;
  multitrmnal?: number;
  consolidated_pathfile?: string;
  refnum?: string;
  recontime?: string;
  sm_mall_2022?: number;
  nexbridge?: number;
  sta_lucia?: number;
  receipt_itmnum?: number;
  shangrila?: number;
  ayala_mall_2022?: number;
  central_lastcon?: string;
  onetime_prcload?: number;
  enablerefund?: number;
  sftp_robinson?: number;
  robinson_operation_set?: number;
  transferlimit?: number;
  viewsentfiles?: number;
  enable_separate_hookup?: number;
  mitsukoshi?: number;
  api_url?: string;
  api_key?: string;
  secret_key?: string;
  request_type?: string;
  contract_no?: string;
  generate_key?: string;
  pos_no?: string;
  comp_name?: string;
  mitsukoshi_setup_type?: string;
  allow_printerstation?: number;
  itemclass_printer_station_tag?: boolean;
  itemsubclass_printer_station_tag?: boolean;
  prntr_stn_tag_modified?: boolean;
  enable_spcl_req_receipt?: boolean;
  receipt_title?: number;
  timeformat?: boolean;
  isextended?: number;
  timeextension?: number;
  timeinterval?: number;
  active_mall?: number;
}